25.06.01 작업 내역
1. 세이브 데이터 완성
   - 빈 슬롯 클릭 - 초기 데이터 생성 - 데이터 셋팅(작업 필요) - 튜토리얼 이동) 순 동작
   - 다만 1번 슬롯이 빈 상태에서 2번 슬롯 데이터를 불러오기 할 경우 에러 발생
   - 데이터를 플레이어에 덮어쓰는 기능 필요함

2. 스테이지 데이터 용 스크립터블 오브젝트 구조 변경
   - 저장 데이터 내에 List<ChapterData> 형태로 보관중이던 데이터를 ClearData 클래스로 묶음
   - 이에 따라 데이터 접근 로직도 ClearData 를 거쳐 접근하도록 변경
   - StageData에 Rank 값 부분 디폴트가 None 인데, N으로 교체 필요 (텍스트가 UI를 벗어남)

3. 아이템 데이터의 경우 기본값은 -1로 저장(아이템이 없을 경우)
   - 아이템 스크립트 노트북에 작성해둔거 이동시킬 준비 할것

4. 스킬 데이터 저장 기능도 필요
   - 스킬 매니저에서 각 스킬별 저장 데이터를 어떻게 구현할지? - class 기반 데이터 고민 중

+++ Assets/Scripts/Player/Player_Manager.cs
@@ -21,7 +21,7 @@ public class Player_Manager : MonoBehaviour, IDamageSysteam
     // Attack Status
     public int physicalDamage;
     public int magicalDamage;
-    public int attackSpeed;
+    public float attackSpeed;
     public float criticalhit;
     public float critical_multiplier;
